Create g_Weather_Details_Application.py
Program to find all the details of weather of a place given by user by using api in the application and other details.
"This code is an example of various libraries taught by seniors in college. A lot to learn and lot to improve in this code and I will be improving the code or perhaps writing a new code from start" later.